Revert "dao/vo, mybatis, log4j, mappers commit"
This reverts commit e786679e084f6c922a062c5b713e71b92a488559.
