+++ src/main/java/gayo/spring/admin/classDAOMybatis.java
@@ -6,7 +6,7 @@ import org.mybatis.spring.SqlSessionTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
-import gayo.spring.admin.vo.TermsVO;
+import gayo.spring.web.admin.vo.TermsVO;
 
 @Repository
 public class classDAOMybatis {
