+++ src/main/java/gayo/spring/web/admin/dao/CustomerDAO.java
@@ -1,11 +1,11 @@
-package gayo.spring.admin.dao;
+package gayo.spring.web.admin.dao;
 
 import java.util.List;
 
 import org.apache.ibatis.session.SqlSession;
 
 import gayo.spring.admin.sqlSessionFactoryBean;
-import gayo.spring.admin.vo.CustomerVO;
+import gayo.spring.web.admin.vo.CustomerVO;
 
 public class CustomerDAO {
 	private SqlSession sqlSession;
