+++ src/main/java/gayo/spring/web/admin/dao/NoticeDAO.java
@@ -1,11 +1,11 @@
-package gayo.spring.admin.dao;
+package gayo.spring.web.admin.dao;
 
 import java.util.List;
 
 import org.apache.ibatis.session.SqlSession;
 
 import gayo.spring.admin.sqlSessionFactoryBean;
-import gayo.spring.admin.vo.NoticeVO;
+import gayo.spring.web.admin.vo.NoticeVO;
 
 public class NoticeDAO {
 	private SqlSession sqlSession;
