+++ src/main/java/gayo/spring/web/admin/dao/QnaDAO.java
@@ -1,11 +1,11 @@
-package gayo.spring.admin.dao;
+package gayo.spring.web.admin.dao;
 
 import java.util.List;
 
 import org.apache.ibatis.session.SqlSession;
 
 import gayo.spring.admin.sqlSessionFactoryBean;
-import gayo.spring.admin.vo.QnaVO;
+import gayo.spring.web.admin.vo.QnaVO;
 
 public class QnaDAO {
 	private SqlSession sqlSession;
