+++ src/main/java/gayo/spring/web/admin/dao/ReserveDAO.java
@@ -1,11 +1,11 @@
-package gayo.spring.admin.dao;
+package gayo.spring.web.admin.dao;
 
 import java.util.List;
 
 import org.apache.ibatis.session.SqlSession;
 
 import gayo.spring.admin.sqlSessionFactoryBean;
-import gayo.spring.admin.vo.ReserveVO;
+import gayo.spring.web.admin.vo.ReserveVO;
 
 public class ReserveDAO {
