+++ src/main/java/gayo/spring/web/admin/dao/UserCountDAO.java
@@ -1,11 +1,11 @@
-package gayo.spring.admin.dao;
+package gayo.spring.web.admin.dao;
 
 import java.util.List;
 
 import org.apache.ibatis.session.SqlSession;
 
 import gayo.spring.admin.sqlSessionFactoryBean;
-import gayo.spring.admin.vo.UserCountVO;
+import gayo.spring.web.admin.vo.UserCountVO;
 
 public class UserCountDAO {
 	private SqlSession sqlSession;
