+++ src/main/java/gayo/spring/web/admin/vo/AdminVO.java
@@ -1,4 +1,4 @@
-package gayo.spring.admin.vo;
+package gayo.spring.web.admin.vo;
 
 public class AdminVO {
 	private String admins_Id;
