+++ src/main/java/gayo/spring/web/admin/vo/BannerVO.java
@@ -1,4 +1,4 @@
-package gayo.spring.admin.vo;
+package gayo.spring.web.admin.vo;
 
 public class BannerVO {
 	private String banner_No;
