+++ src/main/java/gayo/spring/web/admin/vo/PayVO.java
@@ -1,15 +1,17 @@
-package gayo.spring.admin.vo;
+package gayo.spring.web.admin.vo;
 
-public class SalesVO {
+import java.sql.Date;
+
+public class PayVO {
 	private String customer_Id;
-	private String reserve_No;
 	private String reserve_Pay;
+	private Date pay_Date;
+	
 	
 	
 	@Override
 	public String toString() {
-		return "SalesVO [customer_Id=" + customer_Id + ", reserve_No=" + reserve_No + ", reserve_Pay=" + reserve_Pay
-				+ "]";
+		return "PayVO [customer_Id=" + customer_Id + ", reserve_Pay=" + reserve_Pay + ", pay_Date=" + pay_Date + "]";
 	}
 	public String getCustomer_Id() {
 		return customer_Id;
@@ -17,17 +19,19 @@ public class SalesVO {
 	public void setCustomer_Id(String customer_Id) {
 		this.customer_Id = customer_Id;
 	}
-	public String getReserve_No() {
-		return reserve_No;
-	}
-	public void setReserve_No(String reserve_No) {
-		this.reserve_No = reserve_No;
-	}
 	public String getReserve_Pay() {
 		return reserve_Pay;
 	}
 	public void setReserve_Pay(String reserve_Pay) {
 		this.reserve_Pay = reserve_Pay;
 	}
+	public Date getPay_Date() {
+		return pay_Date;
+	}
+	public void setPay_Date(Date pay_Date) {
+		this.pay_Date = pay_Date;
+	}
+	
+	
 	
 }
