+++ src/main/java/gayo/spring/web/admin/vo/ReserveVO.java
@@ -1,4 +1,4 @@
-package gayo.spring.admin.vo;
+package gayo.spring.web.admin.vo;
 
 public class ReserveVO {
 	private String reserve_No;
