+++ src/main/java/gayo/spring/web/admin/vo/SalesVO.java
@@ -1,17 +1,15 @@
-package gayo.spring.admin.vo;
+package gayo.spring.web.admin.vo;
 
-import java.sql.Date;
-
-public class PayVO {
+public class SalesVO {
 	private String customer_Id;
+	private String reserve_No;
 	private String reserve_Pay;
-	private Date pay_Date;
-	
 	
 	
 	@Override
 	public String toString() {
-		return "PayVO [customer_Id=" + customer_Id + ", reserve_Pay=" + reserve_Pay + ", pay_Date=" + pay_Date + "]";
+		return "SalesVO [customer_Id=" + customer_Id + ", reserve_No=" + reserve_No + ", reserve_Pay=" + reserve_Pay
+				+ "]";
 	}
 	public String getCustomer_Id() {
 		return customer_Id;
@@ -19,19 +17,17 @@ public class PayVO {
 	public void setCustomer_Id(String customer_Id) {
 		this.customer_Id = customer_Id;
 	}
+	public String getReserve_No() {
+		return reserve_No;
+	}
+	public void setReserve_No(String reserve_No) {
+		this.reserve_No = reserve_No;
+	}
 	public String getReserve_Pay() {
 		return reserve_Pay;
 	}
 	public void setReserve_Pay(String reserve_Pay) {
 		this.reserve_Pay = reserve_Pay;
 	}
-	public Date getPay_Date() {
-		return pay_Date;
-	}
-	public void setPay_Date(Date pay_Date) {
-		this.pay_Date = pay_Date;
-	}
-	
-	
 	
 }
